chore: update doc-gen output_dir (#1123)
## Description

The generated docs were moved, but it seems the script to generate them
was not updated to reflect the new location. This PR fixes that.

## Related Issue

Relates to #1119 because at the moment it has docs that shouldn't be
generated in it.

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)
- [ ] 
## Checklist before merging

- [ ] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed

Signed-off-by: catsby <clint@defenseunicorns.com>
Co-authored-by: Micah Nagel <micah.nagel@defenseunicorns.com>