chore: add action to check readiness on renovate PRs before running CI (#1144)
## Description

Adds a new action that checks renovate PR readiness and fails CI before
using the large runners if the PR is not ready. At this point the action
just checks for a `renovate-ready` label, but this could be changed in
the future to look for changes in the PR, etc.

## Related Issue

Related to https://github.com/defenseunicorns/uds-core/issues/168

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Steps to Validate

This functionality was validated on [my
fork](https://github.com/BagelLab/uds-core) due to some of the
complexities around testing it on this repo. A demo/walkthrough can be
provided by appointment.

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed