fix: lint errors on unused caught errors (#1271)
## Description

Resolves some errors that show in pepr format due to unused errors that
we catch. In each case we don't have a need to use the error so I just
dropped the variable.

This also resolves an issue with the latest format command where it
would error on staged ts files. By moving to the `cjs` format of the
lint config we can specify how to use the list of files (and in our case
just not use them).

## Related Issue

N/A

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Steps to Validate

May be worth validating the new lint config by committing misspellings
and poorly formatted files.

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed