chore: update how to scrape metrics (#1378)
## Description

Since the images have changed, so has the way to scrape the Pepr pod
metrics endpoint. This PR updates the docs to show how to scrape in an
airgapped environment based on the four images.

The background is based on this
[comment](https://github.com/defenseunicorns/uds-core/issues/1209#issuecomment-2641314635)
and a recent convo with @mjnagel



## Related Issue

Fixes #
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Steps to Validate
- If this PR introduces new functionality to UDS Core or addresses a
bug, please document the steps to test the changes.

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>
Co-authored-by: Micah Nagel <micah.nagel@defenseunicorns.com>