fix: remove flavor from dev deploy of prom CRDs task (#1419)
## Description

Zarf as of 0.51.0 will validate the flavor actually exists, so this
change drops the `--flavor` (which was not necessary in the first
place). This was primarily used as "documentation" so I made an update
to the comment for clarity.

## Related Issue

Zarf change: https://github.com/zarf-dev/zarf/pull/3597

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Steps to Validate

Run the task/command to validate it works as expect (`zarf dev deploy
src/prometheus-stack --no-progress` should only deploy the CRDs).

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed