feat: client credential registration default (#1482)
## Description
In moving uds-core to ambient we need to rework how we implement the
envoyfilter used to protect path parameters. In addition to that, in the
coming release we will phase out the use of the dynamic client
registration in favor client credential registration. This will improve
security and decrease complexity when working with istio ambient.

This PR will focus on the following:
1. Change the default keycloak client registration strategy from `auto`
to `client_credentials`

## Related Issue

Fixes #1388 

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed

---------

Co-authored-by: Sebastian Łaskawiec <sebastian.laskawiec@defenseunicorns.com>