chore: use built-in sdk for containers (#1530)
## Description

Pepr contains an `sdk` of helper functions that are very well tested. It
comes in the library, and one is called
[containers](https://github.com/defenseunicorns/pepr/blob/db6c4c4b8e341dc50fc7b679b272a8ab67a0ee32/src/sdk/sdk.ts#L15)
which returns all containers in a Pod. The idea is that this allowed
Module Authors to reduce code usage and leverage internal functions.

## Related Issue

Fixes #1533 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Steps to Validate
- If this PR introduces new functionality to UDS Core or addresses a
bug, please document the steps to test the changes.

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>
Co-authored-by: Micah Nagel <micah.nagel@defenseunicorns.com>
Co-authored-by: Chance <139784371+UnicornChance@users.noreply.github.com>