feat: configurable Istio mtlsClientCert (#1553)
## Description
This PR allows a configurable mtlsClientCert to use, but still defaults
to `%DOWNSTREAM_PEER_CERT%`. This allows the user to specify another
value such as a different header (needed if TLS terminates at an AWS ALB
and the cert is now in a different header name)

## Related Issue
Discussed offline, needed for an effort

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Steps to Validate
- If this PR introduces new functionality to UDS Core or addresses a
bug, please document the steps to test the changes.

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed

Co-authored-by: Noah <40781376+noahpb@users.noreply.github.com>
Co-authored-by: Chance <139784371+UnicornChance@users.noreply.github.com>