fix: ensure uniqueness of sso client ids in cluster (#1589)
## Description

This PR uses the `PackageStore` in our validator to ensure that SSO
client IDs are "globally" unique across the entire cluster (in addition
to locally unique in the package CR).

## Related Issue

N/A

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Steps to Validate

Deploy a package CR with the same SSO client ID as another package.
Ensure that it is blocked by the validator.

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed