chore(ci): add iac filter workflow (#1639)
## Description
This PR moves IaC tests to a new workflow, separated from the existing
"Filter" workflow that contained all CI tests for K3D and IaC (EKS,
RKE2, AKS). This change is being introduced after discovering that
having all workflows nested under a single workflow is not optimal when
performing release operations.

The workflow will only trigger when a file pertaining to an IaC workflow
is modified or when a label to trigger an IaC workflow is added to a PR.

Additionally, a new concurrency group is introduced that references the
most recent commit hash, with `cancel-in-progress` set to `false`. This
will prevent orphaned resources from premature interruptions to `tofu
destroy` operations
and avoid state lockout issues. 

## Related Issue

Changes from
[this](https://github.com/defenseunicorns/uds-core/actions/runs/15541956969/job/43758321554)
to
[this](https://github.com/defenseunicorns/uds-core/actions/runs/15563537235/job/43821760910)
for IaC Workflow Triggering.

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Steps to Validate
- If this PR introduces new functionality to UDS Core or addresses a
bug, please document the steps to test the changes.

## Checklist before merging

- [ ] Test, docs, adr added or updated as needed
- [ ] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed