chore: enable linting on test dirs and address eslint/formatting issues (#1728)
## Description

Our linting and formatting (eslint/prettier) was not running on certain
dirs that it really should be running on. This PR fixes the lint
workflow and also addresses existing issues in all the new files added
to this lint workflow.

## Related Issue

N/A

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Steps to Validate
- introduce an unused variable in one of the files under test/vitest
- `uds run lint-check` (this should throw an eslint error)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed