fix: clean up testing methodologies (#1863)
## Description
### Context
Want to make a clear deliniation of unit vs e2e testing. at the moment
both are present in the same areas,
making it confusing. E2E is defined as need a cluster and unit is
defined as standalone test that doesnt
require a cluster. Unit tests should continue to live next to the source
code that its written for and E2E
should be moved to the `test/vitest` directory.

### What has been done
- moved four e2e tests to the `test/vitest/pepr-policies` directory
because they are E2E tests not unit tests:
    - `istio.spec.ts`
    - `network.spec.ts`
    - `storage.spec.ts`
    - `security.spec.ts`
- rework the mocking of `authservice.spec.ts` to not require a cluster
- rework the mocking of `client-secret-sync.spec.ts` to not require a
cluster
- move the `vitest.setup.js` from the root of the project to the
`test/vitest` e2e tests directory and update the `vitest.config.js` at
the root of the project to not use a cluster and the `vitest.config.js`
in the `test/vitest` to use a cluster since its e2e tests specifically.
Also update the root `vitest.config.js` to not run e2e vitest files.
- new readme for explaining where tests go and how to run them
- add folders to test/vitest directory for better organization


## Related Issue

Fixes #1735 

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed