fix: pepr policy unit testing (#1880)
## Description
Previously we had E2E testing of the pepr policies, which meant that
during this PR those tests were moved to the E2E vitest directory.
Because of the complexity of testing functionality inside of a the
`When` `Mutate` and `Validate` we decided to refactor the code to move
the uds-core business logic to its own functions so that we could avoid
needing to mock all of the pepr functions. This creates straightforward
simple unit testing scenarios.

No cluster is required for these tests to run. 47 test files, 704 tests,
complete under 2 seconds.

## Related Issue

Relates to #1863

## Type of change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist
- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed

---------

Co-authored-by: Sebastian Łaskawiec <sebastian.laskawiec@defenseunicorns.com>