fix: dev task workflow (#1931)
## Description

Fixes the dev-setup task and workflow. This
[PR](https://github.com/defenseunicorns/uds-core/pull/1896) introduced a
volume mount to istiod for a secret that is not created until pepr is
up. This updates the dev workflow to create this secret ahead of time.

This also cleans up the dev task by pulling all the manifests we need to
apply into a separate folder called `hack/dev-manifests`.

## Related Issue

N/A

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Steps to Validate
- `uds run dev-identity`

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed