fix: keycloak statefulset spi provider syntax (#2030)
## Description
In a previous keycloak update they updated their spi statefulset
reference syntax, it now needs to be double underscores. At the moment
this is just a warning error that happens when not using the correct
syntax.

`WARNING: The following SPI options are using the legacy format and are
not being treated as build time options. Please use the new format with
the appropriate -- separators to resolve this ambiguity:
kc.spi-x509cert-lookup-provider`

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Steps to Validate
- `uds run test-single-layer --set LAYER=identity-authorization` from
main and monitoring the logs of keycloak pod as it comes up to see the
warning
- run the same command from my branch and see that the warning is no
longer present.

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed